fix: admit changes-requested runtime redispatch
